# docs/openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrailBase",
    "description": "TrailBase APIs",
    "license": {
      "name": "OSL-3.0",
      "identifier": "OSL-3.0"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/auth/v1/change_email/request": {
      "post": {
        "tags": ["auth"],
        "summary": "Request an email change.",
        "operationId": "change_email_request_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success."
          }
        }
      }
    },
    "/api/auth/v1/change_password": {
      "post": {
        "tags": ["auth"],
        "summary": "Request a change of password.",
        "operationId": "change_password_handler",
        "parameters": [
          {
            "name": "redirect_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success."
          }
        }
      }
    },
...
    "/api/auth/v1/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Log in users by email and password.",
        "operationId": "login_handler",
        "parameters": [
          {
            "name": "redirect_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Auth & refresh tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/v1/oauth/{provider}/login": {
      "get": {
        "tags": ["oauth"],
        "summary": "Log in via external OAuth provider.",
        "operationId": "login_with_external_auth_provider",
        "parameters": [
          {
            "name": "redirect_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "name": "response_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "name": "pkce_code_challenge",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Redirect."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeEmailRequest": {
        "type": "object",
        "required": ["csrf_token", "new_email"],
        "properties": {
          "csrf_token": {
            "type": "string"
          },
          "new_email": {
            "type": "string"
          },
          "old_email": {
            "type": ["string", "null"]
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": ["old_password", "new_password", "new_password_repeat"],
        "properties": {
          "new_password": {
            "type": "string"
          },
          "new_password_repeat": {
            "type": "string"
          },
          "old_password": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "pkce_code_challenge": {
            "type": ["string", "null"]
          },
          "redirect_to": {
            "type": ["string", "null"]
          },
          "response_type": {
            "type": ["string", "null"]
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["auth_token", "refresh_token", "csrf_token"],
        "properties": {
          "auth_token": {
            "type": "string"
          },
          "csrf_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}